+++ amplitude.js
@@ -3144,7 +3144,7 @@ module.exports = uuid;
 
 }, {}],
 12: [function(require, module, exports) {
-module.exports = '2.4.0';
+module.exports = '2.4.1';
 
 }, {}]}, {}, {"1":""})
 );