+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 ## Unreleased
 
+* Add support for user properties operations (set, setOnce, add, unset).
+
 ## 2.4.0 (September 4, 2015)
 
-* Add support for passing callback functions to logEvent
+* Add support for passing callback functions to logEvent.
 
 ## 2.3.0 (September 2, 2015)
 
@@ -11,7 +13,7 @@
 ## 2.2.1 (Aug 13, 2015)
 
 * Fix bug where multi-byte unicode characters were hashed improperly.
-* Add option to sent referrer information as user properties
+* Add option to send referrer information as user properties.
 
 ## 2.2.0 (May 5, 2015)